Fixed a lot in this one.
We updated functions so that it uses normal gw values because we were getting an index error when looking for key of gw scores.

We found a very rare (maybe not so much) occasion that all fixtures were cancelled for a gw due to the death of the Queen so everyone's score just ended up as 0. This meant that it worked to find lowest score but to find nth lowest was returning an empty list of names because "Nobody" was nth lowest. This threw up an index error again because it was trying to use an empty list to show who was nth lowest for a game week.

Added "Nobody" to the report to show that nobody was nth lowest for the comments. This annoyingly then created a row for "Nobody" in the dataframe which is why I had to filter that out in routes.py just before I render template.

Likewise, this also added it to the breakdown at the bottom showing how much people owed and gave "Nobody" an amount. Also filtered this out using Jinja in the fpl.html template.

Finally, I made it so that the breakdown buttons for 2nd and 3rd bottom only show depending on which form option you use. 1 will only show the "show breakdown" for lowest 2 will show 2nd lowest and lowest and 3 will show all 3.